# Greatly improve windows performance