Fixes #216 

@tkozybski: Your help was really appreciated on this, thanks a lot for pointing me to this API! A few numbers:

```
=> before:

running 8 tests
test bench_new                ... bench: 597,681,058 ns/iter (+/- 520,389,224)
test bench_refresh_all        ... bench: 192,556,222 ns/iter (+/- 107,715,769)
test bench_refresh_disk_lists ... bench:     345,887 ns/iter (+/- 29,159)
test bench_refresh_disks      ... bench:     112,346 ns/iter (+/- 15,075)
test bench_refresh_network    ... bench:         203 ns/iter (+/- 14)
test bench_refresh_process    ... bench:       2,706 ns/iter (+/- 176)
test bench_refresh_processes  ... bench: 191,591,762 ns/iter (+/- 5,917,054)
test bench_refresh_system     ... bench:       2,489 ns/iter (+/- 456)

=> after:

running 8 tests
test bench_new                ... bench:  48,019,473 ns/iter (+/- 210,328,340)
test bench_refresh_all        ... bench:   1,984,283 ns/iter (+/- 258,629)
test bench_refresh_disk_lists ... bench:     342,007 ns/iter (+/- 46,104)
test bench_refresh_disks      ... bench:     113,166 ns/iter (+/- 10,406)
test bench_refresh_network    ... bench:         207 ns/iter (+/- 22)
test bench_refresh_process    ... bench:       2,730 ns/iter (+/- 213)
test bench_refresh_processes  ... bench:   1,755,207 ns/iter (+/- 150,654)
test bench_refresh_system     ... bench:       2,546 ns/iter (+/- 929)
```