@tkozybski 2019-12-07
@GuillaumeGomez  Great. For getting the process name, I suggest this:
```
        let name = &process.ImageName;
        let proc_name = if name.Buffer.is_null() {
            match process_id {
                0 => "Idle".to_string(),
                4 => "System".to_string(),
                _ => process_id.to_string(),
            }
        } else {
            let slice = unsafe {
                std::slice::from_raw_parts(
                    name.Buffer,
                    //The length is in bytes, not the length of string
                    name.Length as usize / std::mem::size_of::<u16>(),
                )
            };

            String::from_utf16_lossy(slice)
        };
```

I'm not sure if the slice should remove the null termination character before calling from_utf16_lossy.