@GuillaumeGomez 2019-12-07
@tkozybski Thanks a lot once again! You should definitely try to send a PR in the future so your contributions can be seen... I'm sending a PR with your suggested changes.